Add files via upload
Shiny UI Demo. Reacts to which dataset is chosen. I will also upload the paired down version of our code. We can use this as the structure for our project. The output changes based on user selection. If we can just construct datasets with the appropriate variable situations and make the plot command "reactive" we should be in business.